+++ app/src/main/java/com/example/returnpals/dataRepository/CognitoMainMenuScreenRepository.kt
@@ -1,4 +1,4 @@
-package com.example.returnpals.data
+package com.example.returnpals.dataRepository
 
 
 import com.amplifyframework.core.Amplify
