+++ app/src/main/java/com/example/returnpals/dataRepository/CognitoRegisterRepository.kt
@@ -1,4 +1,4 @@
-package com.example.returnpals.data
+package com.example.returnpals.dataRepository
 
 import android.util.Log
 import com.amplifyframework.api.graphql.model.ModelMutation
