+++ app/src/main/java/com/example/returnpals/dataRepository/MainMenuScreenRepository.kt
@@ -1,4 +1,4 @@
-package com.example.returnpals.data
+package com.example.returnpals.dataRepository
 
 
 //This file exposes the call to see if the user is logged in for navigating away from main menu
