+++ app/src/main/java/com/example/returnpals/navigation/MainMenu.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavController
 import com.example.returnpals.composetools.mainMenu.MainMenuScreen
 import com.example.returnpals.composetools.mainMenu.MainMenuScreenViewModel
-import com.example.returnpals.data.CognitoMainMenuScreenRepository
+import com.example.returnpals.dataRepository.CognitoMainMenuScreenRepository
 
 //This file create the navigation for the main menu.
 //Then navigate to either the dashboard or sign in.
